requirements.txt
python-fasthtml
pandas
python-dateutil 
redis
